requirements.txt
fastapi
uvicorn
google-generativeai
requests
python-dotenv
jinja2
python-multipart
httpx